+++ packages/amplify-category-function/src/provider-utils/awscloudformation/utils/buildFunction.ts
@@ -1,14 +1,14 @@
 import { $TSContext, pathManager } from 'amplify-cli-core';
 import { FunctionRuntimeLifecycleManager, BuildRequest, BuildType } from 'amplify-function-plugin-interface';
 import * as path from 'path';
-import { category } from '../../../constants';
+import { categoryName } from '../../../constants';
 
 export const buildFunction = async (
   context: $TSContext,
-  { resourceName, lastBuildTimeStamp, buildType = BuildType.PROD }: BuildRequestMeta,
+  { resourceName, lastBuildTimestamp, buildType = BuildType.PROD }: BuildRequestMeta,
 ) => {
-  const resourcePath = path.join(pathManager.getBackendDirPath(), category, resourceName);
-  const breadcrumbs = context.amplify.readBreadcrumbs(category, resourceName);
+  const resourcePath = path.join(pathManager.getBackendDirPath(), categoryName, resourceName);
+  const breadcrumbs = context.amplify.readBreadcrumbs(categoryName, resourceName);
 
   const runtimePlugin: FunctionRuntimeLifecycleManager = (await context.amplify.loadRuntimePlugin(
     context,
@@ -21,7 +21,7 @@ export const buildFunction = async (
     throw new Error(`Missing required dependencies to package ${resourceName}`);
   }
 
-  const prevBuildTime = lastBuildTimeStamp ? new Date(lastBuildTimeStamp) : undefined;
+  const prevBuildTime = lastBuildTimestamp ? new Date(lastBuildTimestamp) : undefined;
 
   // build the function
   let rebuilt = false;
@@ -42,16 +42,16 @@ export const buildFunction = async (
     rebuilt = (await runtimePlugin.build(buildRequest)).rebuilt;
   }
   if (rebuilt) {
-    context.amplify.updateamplifyMetaAfterBuild({ category, resourceName }, buildType.toString());
+    context.amplify.updateamplifyMetaAfterBuild({ category: categoryName, resourceName }, buildType.toString());
     return new Date().toISOString();
   } else {
-    return lastBuildTimeStamp;
+    return lastBuildTimestamp;
   }
 };
 
 export interface BuildRequestMeta {
   resourceName: string;
-  lastBuildTimeStamp?: string;
+  lastBuildTimestamp?: string;
   buildType?: BuildType;
 }