+++ packages/amplify-category-function/src/provider-utils/awscloudformation/utils/getDependentFunction.ts
@@ -1,8 +1,8 @@
 import { $TSContext, $TSObject } from 'amplify-cli-core';
+import * as path from 'path';
 import { loadFunctionParameters } from './loadFunctionParameters';
-import path from 'path';
 import { ServiceName } from './constants';
-import { category } from '../../../constants';
+import { categoryName } from '../../../constants';
 
 export async function lambdasWithApiDependency(
   context: $TSContext,
@@ -22,8 +22,8 @@ export async function lambdasWithApiDependency(
 
   // initialize function parameters for update
   for (const lambda of lambdaFuncResources) {
-    const resourceDirPath = path.join(backendDir, category, lambda.resourceName);
-    const currentParameters = loadFunctionParameters(context, resourceDirPath);
+    const resourceDirPath = path.join(backendDir, categoryName, lambda.resourceName);
+    const currentParameters = loadFunctionParameters(resourceDirPath);
     const selectedCategories = currentParameters.permissions;
     let deletedModelFound: boolean;