+++ packages/amplify-cli/src/index.ts
@@ -1,15 +1,15 @@
 import {
   $TSAny,
   $TSContext,
+  BannerMessage,
   CLIContextEnvironmentProvider,
   exitOnNextTick,
   FeatureFlags,
+  JSONUtilities,
   JSONValidationError,
   pathManager,
   stateManager,
   TeamProviderInfoMigrateError,
-  BannerMessage,
-  JSONUtilities,
 } from 'amplify-cli-core';
 import { isCI } from 'ci-info';
 import { EventEmitter } from 'events';
@@ -17,7 +17,8 @@ import * as fs from 'fs-extra';
 import * as path from 'path';
 import { logInput } from './conditional-local-logging-init';
 import { print } from './context-extensions';
-import { attachUsageData, constructContext, persistContext } from './context-manager';
+import { attachUsageData, constructContext } from './context-manager';
+import { displayBannerMessages } from './display-banner-messages';
 import { constants } from './domain/constants';
 import { Context } from './domain/context';
 import { Input } from './domain/input';
@@ -40,7 +41,7 @@ Error.stackTraceLimit = Number.MAX_SAFE_INTEGER;
 
 let errorHandler = (e: Error) => {};
 
-process.on('uncaughtException', function(error) {
+process.on('uncaughtException', function (error) {
   // Invoke the configured error handler if it is already configured
   if (errorHandler) {
     errorHandler(error);
@@ -59,7 +60,7 @@ process.on('uncaughtException', function(error) {
 });
 
 // In this handler we have to rethrow the error otherwise the process stucks there.
-process.on('unhandledRejection', function(error) {
+process.on('unhandledRejection', function (error) {
   throw error;
 });
 
@@ -70,6 +71,7 @@ export async function run() {
 
     let pluginPlatform = await getPluginPlatform();
     let input = getCommandLineInput(pluginPlatform);
+
     // with non-help command supplied, give notification before execution
     if (input.command !== 'help') {
       // Checks for available update, defaults to a 1 day interval for notification
@@ -143,6 +145,9 @@ export async function run() {
       return 1;
     }
 
+    // Display messages meant for most executions
+    await displayBannerMessages(input);
+
     await executeCommand(context);
 
     const exitCode = process.exitCode || 0;
@@ -150,8 +155,6 @@ export async function run() {
     if (exitCode === 0) {
       context.usageData.emitSuccess();
     }
-
-    persistContext(context);
 
     // no command supplied defaults to help, give update notification at end of execution
     if (input.command === 'help') {
@@ -284,8 +287,6 @@ export async function execute(input: Input): Promise<number> {
     if (exitCode === 0) {
       context.usageData.emitSuccess();
     }
-
-    persistContext(context);
 
     return exitCode;
   } catch (e) {