+++ packages/graphql-transformers-e2e-tests/src/__tests__/ConnectionsWithAuthTests.e2e.test.ts
@@ -29,6 +29,17 @@ import 'isomorphic-fetch';
 jest.setTimeout(2000000);
 
 const cf = new CloudFormationClient('us-west-2');
+const featureFlags = {
+  getBoolean: jest.fn().mockImplementation((name, defaultValue) => {
+    if (name === 'improvePluralization') {
+      return true;
+    }
+    return;
+  }),
+  getNumber: jest.fn(),
+  getObject: jest.fn(),
+  getString: jest.fn(),
+};
 
 const BUILD_TIMESTAMP = moment().format('YYYYMMDDHHmmss');
 const STACK_NAME = `ConnectionsWithAuthTests-${BUILD_TIMESTAMP}`;
@@ -148,6 +159,7 @@ beforeAll(async () => {
     }
     `;
   const transformer = new GraphQLTransform({
+    featureFlags,
     transformers: [
       new DynamoDBModelTransformer(),
       new ModelConnectionTransformer(),