+++ packages/graphql-transformers-e2e-tests/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.24.1](https://github.com/aws-amplify/amplify-cli/compare/graphql-transformers-e2e-tests@6.24.0...graphql-transformers-e2e-tests@6.24.1) (2021-06-15)
+
+**Note:** Version bump only for package graphql-transformers-e2e-tests
+
+
+
+
+
+# [6.24.0](https://github.com/aws-amplify/amplify-cli/compare/graphql-transformers-e2e-tests@6.23.10...graphql-transformers-e2e-tests@6.24.0) (2021-06-02)
+
+
+
+# 4.52.0 (2021-06-01)
+
+
+### Features
+
+* add support for SMS Sandbox ([#7436](https://github.com/aws-amplify/amplify-cli/issues/7436)) ([cdcb626](https://github.com/aws-amplify/amplify-cli/commit/cdcb6260c11bbedef5b056fdcd730612d8bb3230))
+
+
+
+
+
+## [6.23.10](https://github.com/aws-amplify/amplify-cli/compare/graphql-transformers-e2e-tests@6.23.9...graphql-transformers-e2e-tests@6.23.10) (2021-05-29)
+
+
+
+## 4.51.4 (2021-05-28)
+
+
+### Bug Fixes
+
+* **graphql-model-transformer:** use modelobject key for mutation resolver creation ([#7419](https://github.com/aws-amplify/amplify-cli/issues/7419)) ([37bc551](https://github.com/aws-amplify/amplify-cli/commit/37bc551030d47de993f8227ee3af0ba6cd738ab2)), closes [#i7417](https://github.com/aws-amplify/amplify-cli/issues/i7417)
+
+
+
+
+
+## [6.23.9](https://github.com/aws-amplify/amplify-cli/compare/graphql-transformers-e2e-tests@6.23.8...graphql-transformers-e2e-tests@6.23.9) (2021-05-26)
+
+**Note:** Version bump only for package graphql-transformers-e2e-tests
+
+
+
+
+
 ## [6.23.8](https://github.com/aws-amplify/amplify-cli/compare/graphql-transformers-e2e-tests@6.23.7...graphql-transformers-e2e-tests@6.23.8) (2021-05-18)
 
 **Note:** Version bump only for package graphql-transformers-e2e-tests